refactor(admin): extract community config hooks to separate file
- Create new useCommunityConfig.ts hook file with CommunityConfig interface
- Move useCommunityConfig and useCommunityConfigMutation hooks to dedicated file
- Remove 67 lines of hook definitions from Admin component
- Clean up React Query imports and unused mutation instance
- Remove debug console.log statements for cleaner production code
- Improve separation of concerns and code maintainability
